fix(common/models): relocates sortNode to avoid cyclic imports
